+++ Example/nrf-mesh/app/src/main/java/no/nordicsemi/android/nrfmeshprovisioner/node/PublicationSettingsActivity.java
@@ -34,12 +34,12 @@ import no.nordicsemi.android.meshprovisioner.utils.MeshParserUtils;
 import no.nordicsemi.android.meshprovisioner.utils.PublicationSettings;
 import no.nordicsemi.android.nrfmeshprovisioner.R;
 import no.nordicsemi.android.nrfmeshprovisioner.di.Injectable;
-import no.nordicsemi.android.nrfmeshprovisioner.dialog.DialogFragmentPubRetransmitIntervalSteps;
-import no.nordicsemi.android.nrfmeshprovisioner.dialog.DialogFragmentPublicationResolution;
-import no.nordicsemi.android.nrfmeshprovisioner.dialog.DialogFragmentPublicationSteps;
-import no.nordicsemi.android.nrfmeshprovisioner.dialog.DialogFragmentPublishAddress;
-import no.nordicsemi.android.nrfmeshprovisioner.dialog.DialogFragmentPublishTtl;
-import no.nordicsemi.android.nrfmeshprovisioner.dialog.DialogFragmentRetransmitCount;
+import no.nordicsemi.android.nrfmeshprovisioner.node.dialog.DialogFragmentPubRetransmitIntervalSteps;
+import no.nordicsemi.android.nrfmeshprovisioner.node.dialog.DialogFragmentPublicationResolution;
+import no.nordicsemi.android.nrfmeshprovisioner.node.dialog.DialogFragmentPublicationSteps;
+import no.nordicsemi.android.nrfmeshprovisioner.node.dialog.DialogFragmentPublishAddress;
+import no.nordicsemi.android.nrfmeshprovisioner.node.dialog.DialogFragmentPublishTtl;
+import no.nordicsemi.android.nrfmeshprovisioner.node.dialog.DialogFragmentRetransmitCount;
 import no.nordicsemi.android.nrfmeshprovisioner.keys.AppKeysActivity;
 import no.nordicsemi.android.nrfmeshprovisioner.utils.Utils;
 import no.nordicsemi.android.nrfmeshprovisioner.viewmodels.PublicationViewModel;
