+++ Example/nrf-mesh/app/src/main/java/no/nordicsemi/android/nrfmeshprovisioner/node/dialog/DialogFragmentPublicationResolution.java
@@ -1,4 +1,4 @@
-package no.nordicsemi.android.nrfmeshprovisioner.dialog;
+package no.nordicsemi.android.nrfmeshprovisioner.node.dialog;
 
 
 import androidx.appcompat.app.AlertDialog;
