+++ Example/nrf-mesh/app/src/main/java/no/nordicsemi/android/nrfmeshprovisioner/node/dialog/DialogFragmentPublishTtl.java
@@ -20,7 +20,7 @@
  * USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package no.nordicsemi.android.nrfmeshprovisioner.dialog;
+package no.nordicsemi.android.nrfmeshprovisioner.node.dialog;
 
 import androidx.appcompat.app.AlertDialog;
 
@@ -39,56 +39,56 @@ import android.text.TextUtils;
 import android.text.TextWatcher;
 import android.view.LayoutInflater;
 import android.view.View;
-import android.widget.TextView;
+import android.widget.CheckBox;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import no.nordicsemi.android.meshprovisioner.utils.MeshParserUtils;
 import no.nordicsemi.android.nrfmeshprovisioner.R;
 
+public class DialogFragmentPublishTtl extends DialogFragment {
 
-public class DialogFragmentRetransmitCount extends DialogFragment {
-
-    private static final String RETRANSMIT_COUNT = "RETRANSMIT_COUNT";
+    private static final String PUBLISH_TTL = "PUBLISH_TTL";
     //UI Bindings
+    @BindView(R.id.chk_default_ttl)
+    CheckBox chkPublishTtl;
     @BindView(R.id.text_input_layout)
-    TextInputLayout retransmitCountInputLayout;
+    TextInputLayout ttlInputLayout;
     @BindView(R.id.text_input)
-    TextInputEditText retransmitInput;
+    TextInputEditText ttlInput;
 
-    private int mRetransmitCount;
+    private int mPulishTtl;
 
-    public static DialogFragmentRetransmitCount newInstance(final int ivIndex) {
-        DialogFragmentRetransmitCount fragmentIvIndex = new DialogFragmentRetransmitCount();
+    public static DialogFragmentPublishTtl newInstance(final int ttl) {
+        DialogFragmentPublishTtl fragmentNetworkKey = new DialogFragmentPublishTtl();
         final Bundle args = new Bundle();
-        args.putInt(RETRANSMIT_COUNT, ivIndex);
-        fragmentIvIndex.setArguments(args);
-        return fragmentIvIndex;
+        args.putInt(PUBLISH_TTL, ttl);
+        fragmentNetworkKey.setArguments(args);
+        return fragmentNetworkKey;
     }
 
     @Override
     public void onCreate(@Nullable final Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         if (getArguments() != null) {
-            mRetransmitCount = getArguments().getInt(RETRANSMIT_COUNT);
+            mPulishTtl = getArguments().getInt(PUBLISH_TTL);
         }
     }
 
     @NonNull
     @Override
     public Dialog onCreateDialog(final Bundle savedInstanceState) {
-        @SuppressLint("InflateParams") final View rootView = LayoutInflater.from(getContext()).inflate(R.layout.dialog_fragment_publication_parameters, null);
-
-        //Bind ui
+        @SuppressLint("InflateParams") final View rootView = LayoutInflater.from(getContext()).inflate(R.layout.dialog_fragment_publish_ttl_input, null);
         ButterKnife.bind(this, rootView);
-        ((TextView)rootView.findViewById(R.id.summary)).setText(R.string.dialog_summary_retransmit_count);
 
-        final String ivIndex = String.valueOf(mRetransmitCount);
-        retransmitInput.setInputType(InputType.TYPE_CLASS_NUMBER);
-        retransmitCountInputLayout.setHint(getString(R.string.hint_retransmit_count));
-        retransmitInput.setText(ivIndex);
-        retransmitInput.setSelection(ivIndex.length());
-        retransmitInput.addTextChangedListener(new TextWatcher() {
+        chkPublishTtl.setOnCheckedChangeListener((buttonView, isChecked) -> {
+            ttlInputLayout.setVisibility(isChecked ? View.GONE : View.VISIBLE);
+            if(isChecked){
+                ttlInputLayout.setError(null);
+            }
+        });
+
+        ttlInput.addTextChangedListener(new TextWatcher() {
             @Override
             public void beforeTextChanged(final CharSequence s, final int start, final int count, final int after) {
 
@@ -97,9 +97,9 @@ public class DialogFragmentRetransmitCount extends DialogFragment {
             @Override
             public void onTextChanged(final CharSequence s, final int start, final int before, final int count) {
                 if (TextUtils.isEmpty(s.toString())) {
-                    retransmitCountInputLayout.setError(getString(R.string.error_empty_pub_retransmit_count));
+                    ttlInputLayout.setError(getString(R.string.error_empty_publish_ttl));
                 } else {
-                    retransmitCountInputLayout.setError(null);
+                    ttlInputLayout.setError(null);
                 }
             }
 
@@ -112,50 +112,62 @@ public class DialogFragmentRetransmitCount extends DialogFragment {
         final AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(requireContext()).setView(rootView)
                 .setPositiveButton(R.string.ok, null).setNegativeButton(R.string.cancel, null);
 
-        alertDialogBuilder.setIcon(R.drawable.ic_numeric);
-        alertDialogBuilder.setTitle(R.string.title_retransmit_count);
+        alertDialogBuilder.setIcon(R.drawable.ic_timer);
+        alertDialogBuilder.setTitle(R.string.title_publish_ttl);
 
         final AlertDialog alertDialog = alertDialogBuilder.show();
         alertDialog.getButton(DialogInterface.BUTTON_POSITIVE).setOnClickListener(v -> {
-            final String ivIndexInput = this.retransmitInput.getEditableText().toString();
-            if (validateInput(ivIndexInput)) {
-                if (getParentFragment() == null) {
-                    ((DialogFragmentRetransmitCountListener) requireActivity()).setRetransmitCount(Integer.parseInt(ivIndexInput, 16));
-                } else {
-                    ((DialogFragmentRetransmitCountListener) getParentFragment()).setRetransmitCount(Integer.parseInt(ivIndexInput, 16));
-                }
+            if(chkPublishTtl.isChecked()) {
+                ((DialogFragmentPublishTtlListener) requireActivity()).setPublishTtl(MeshParserUtils.USE_DEFAULT_TTL);
                 dismiss();
+            } else {
+                final String publishTtl = ttlInput.getEditableText().toString();
+                if (validateInput(publishTtl)) {
+                    ((DialogFragmentPublishTtlListener) requireActivity()).setPublishTtl(Integer.parseInt(publishTtl));
+                    dismiss();
+                }
             }
         });
+
+        if(savedInstanceState == null) {
+
+            //Update ui
+            if(MeshParserUtils.isDefaultPublishTtl(mPulishTtl)){
+                chkPublishTtl.setChecked(true);
+            } else {
+                final String ttl = String.valueOf(mPulishTtl);
+                ttlInput.setInputType(InputType.TYPE_CLASS_NUMBER);
+                ttlInput.setText(ttl);
+                ttlInput.setSelection(ttl.length());
+            }
+            ttlInputLayout.setHint(getString(R.string.hint_publish_ttl));
+        }
 
         return alertDialog;
     }
 
     private boolean validateInput(final String input) {
-
         try {
-
-            if(TextUtils.isEmpty(input)) {
-                retransmitCountInputLayout.setError(getString(R.string.error_empty_pub_retransmit_count));
+            if(TextUtils.isEmpty(input)){
+                ttlInputLayout.setError(getString(R.string.error_empty_publish_ttl));
                 return false;
             }
-            if (!MeshParserUtils.validateRetransmitCount(Integer.valueOf(input))) {
-                retransmitCountInputLayout.setError(getString(R.string.error_invalid_pub_retransmit_count));
+            if(!MeshParserUtils.isValidTtl(Integer.parseInt(input))) {
+                ttlInputLayout.setError(getString(R.string.error_invalid_publish_ttl));
                 return false;
             }
-        } catch (NumberFormatException ex) {
-            retransmitCountInputLayout.setError(getString(R.string.error_invalid_pub_retransmit_count));
+        } catch(NumberFormatException ex) {
+            ttlInputLayout.setError(getString(R.string.error_invalid_publish_ttl));
             return false;
         } catch (Exception ex) {
             return false;
         }
-
         return true;
     }
 
-    public interface DialogFragmentRetransmitCountListener {
+    public interface DialogFragmentPublishTtlListener {
 
-        void setRetransmitCount(final int retransmitCount);
+        void setPublishTtl(final int ttl);
 
     }
 }
